refactor(cli): migrate worldpop tooling onto discovery (land theme complete)
Move worldpop's refresher/writer/curated_ids/prober/validator to
earthlens.worldpop.cli and drop them from core. Relocate its CLI tests to
earthlens-land. land is the third theme with zero backends named in core; the
guard allow-lists are down to imagery + atmosphere.
